Create task-trackin-service.properties
This is a microservice created for the Rutger Java Boot Camp U1-M1 summitive assessment. This properties file will be used by the tracker service to connect to the server once the config server is up an running. In conjunction with a Eureka Registry, this service will connect to a database to store, update, read, and delete tasks on a database. Additionally, uring Eureka we will implement a way to connect to another "Advertisment" service which will be used in our service.